Pushed XForms to RTSkeleton.  Added global up axis rotation.
The dual skeleton hierarchy in RTClient has been replaced with a single skeleton hierarchy that simply keeps the raw position and rotation data from QTM.  The specific transforms for Mecanim or custom skeletons are now handled in RTSkeleton.

The QTM up axis value is used to determine a global rotation for the data.  Though a new method was created in Rotation to handle this, it is much simpler than the existing one which I couldn't figure out, so I skipped it.